Fix default empty file name
If a user sets to add a file name to the output URL and after sanitizing all characters are removed, the "noroot" name leads to downloading the file with the original name, which may cause issues on the user's system. Using the "file" name instead will lead to downloading the file with the name "file.[ext]".